In short: added a predetermined favorite genres (see why in extended) checklist at the start of the app which is stored as shared prefences upon signing in.
extended:

getGenres didn't work for me, so I manually created a list of genres.

Now, assuming that, at the start of the application, assuming it's your first time, there's a list of genres and the Sign In button, you can check the Genres you like and press Sign in, after that, the list of genres is saved in shared preferences and the UI is updated to only a log out button, if you kill the app now and restart in it'll stay connected (show the log out button) and won't show the genre list, as it's it's already saved in shared preferences.

Signing out won't reset shared prefences as it's saved locally and has nothing to do with google, it's only saved upon pressing the sign in button.